Compress_experiments/neurips_compDFT/requirements.txt
pytorch-pfn-extras
scikit-learn
einops
tqdm
pyyaml
matplotlib
